requirements.txt
omegaconf
rich
torch
torchmetrics
torchvision
pytorch-lightning
hydra-core
wandb
pyrootutils